docker test/requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
pyDes==2.0.1
pytz==2020.4
sqlparse==0.4.1
python-dateutil
parsedatetime
django-cors-headers
